Improve controls abstraction
